// src/enums/NotificationType.ts
export enum NotificationTypeEnum {
  OBJECTIVE_EXPIRED = 'objectiveExpired',
  OBJECTIVE_COMPLETED = 'objectiveCompleted',
  OBJECTIVE_HALF_COMPLETED = 'objectiveHalfCompleted',
  OBJECTIVE_FINISHED = 'objectiveFinished',
  OBJECTIVE_EXPIRING_IN_FIVE_DAYS = 'objectiveExpiringInFiveDays',
  OBJECTIVE_EXPIRING_TOMORROW = 'objectiveExpiringTomorrow',
  OPERATION_FIVE_HUNDRED_OUTFLOW = 'operationFiveHundredOutflow',
  OPERATION_ONE_THOUSAND_OUTFLOW = 'operationOneThousandOutflow',
  OPERATION_TWO_THOUSAND_OUTFLOW = 'operationTwoThousandOutflow',
  OPERATION_THREE_THOUSAND_OUTFLOW = 'operationThreeThousandOutflow',
  OPERATION_FIVE_THOUSAND_OUTFLOW = 'operatioFiveThousandOutflow',
  EFFECT_OPERATION = 'effectOperation',
  CATEGORY_BEST_RESULTS = 'categoryBestResults',
  CATEGORY_WORST_RESULTS = 'categoryWorstResults',
  CATEGORY_LIMIT_EXCEEDED = 'categoryLimitExceeded'
}